RandomQuote: first of the Ten Models
Implement a simple model that gets random quotes from the module
source's <DATA>.  Instead of just emitting lines and list markup from
the controller, emit JSON responses instead so one can use a UI from the
index page; this turns it more of an exercise not only of Perl, but of
HTML5 and JavaScript as well.

Uses jQuery for retrieving quotes from the app, and Modernizr for
enhancing the CSS, with Silkscreen font from Font Squirrel.