Use local copies of jQuery and Modernizr
Grab the latest development version of Modernizr and put it in the
source so we can refer to it locally, instead of calling over remote
HTTP.  With jQuery already bundled in Mojolicious, it makes more sense
for now to use local copies of the JavaScript libraries directly instead
of using them as fallbacks.  I can always change this later.