Add fallback result page for HTML (no JavaScript,) prettify some more
Mojolicious::Controller recently got a respond_to() method allowing for
some nifty content negotiation; use that to turn /moonphase/check into a
RESTful action that returns results in either JSON or HTML.  A little
reading of its magic led me to the horrible state of the the Accept HTTP
request header, so instead of relying on that, use a hidden 'format'
stash value when submitting request via jQuery.ajax().

Also, prettify the result box: it is now a <div>, not a <section>, and
give it the same CSS effects as the enclosing article.