Add controller test for CoverImages
Unlike the model test for Goodreads, we do not have a concern for
needing a developer key or making a real request for book cover images;
we simply need to use the model and expect certain outputs.  For this,
we use Test::MockModule to mock `covers_for_title` as needed in our
test.