Add require.js into the frontend mix
Its 2018, I probably should update this to use React or Vue, but meh,
let's try to upgrade this slowly.  Looking to move to KnockoutJS first,
then perhaps go to Vue.

I also want to avoid using npm/yarn/grunt/gulp/whathaveyou as the focus
is on the Perl backend models, not on the frontend.