:boxing-glove: KO moonphase
This needed some slight updating of the CSS to get the result and error
sections styled right.  Also serialize using `ko.toJSON`so that
jQuery.form would be no longer used.

Result and error sections are in an inline HTML template; these should
probably go in as KO components, maybe for later...